Move the practice widgets to the widgets directory to minimize extra directories. Don't add them to the widget list.
